Private function and compile warnings.

+++ NuoWindow/NuoWindow/AppAboutDialog.cpp
@@ -18,14 +18,15 @@ void AppAboutDialog::InitDialog()
 	NuoDialog::InitDialog();
 
 	float dpi = DPI();
-	NuoRect<long> pos(20 * dpi, 20 * dpi, 98 * dpi, 15 * dpi);
+	NuoRect<long> pos((long)(20 * dpi), (long)(20 * dpi),
+					  (long)(98 * dpi), (long)(15 * dpi));
 
 	_label = std::make_shared<NuoLabel>(shared_from_this());
 	_label->Init();
 	_label->SetText("Nuo Window");
 	_label->SetPositionDevice(pos, false);
 
-	pos.SetY(pos.Y() + pos.H() + 6 * dpi);
+	pos.SetY((long)(pos.Y() + pos.H() + 6 * dpi));
 	_labelCopyright = std::make_shared<NuoLabel>(shared_from_this());
 	_labelCopyright->Init();
 	_labelCopyright->SetText("Copyright (c) 2020");
